+++ fhe/BigTorusPolynomial.cpp
@@ -187,7 +187,7 @@ void copy(BigTorusPolynomial &out, const BigTorusPolynomial &in) {
 
 void precise_conv_toBigReal(BigReal *array, const BigTorusPolynomial &b, int64_t msb) {
 
-    for (int i = 0; i < b.length; i++) {
+    for (uint64_t i = 0; i < b.length; i++) {
         precise_conv_toBigReal(array[i], b.getAT(i), msb);
     }
 
@@ -195,7 +195,7 @@ void precise_conv_toBigReal(BigReal *array, const BigTorusPolynomial &b, int64_t
 
 void BigRealPoly_shift_toBigTorus(BigTorusPolynomial &out, BigReal *array, int64_t left_shift) {
 
-    for (int i = 0; i < out.length; i++) {
+    for (uint64_t i = 0; i < out.length; i++) {
         shift_toBigTorus(out.getAT(i), array[i], left_shift);
     }
 }
